reconcile function fixed

+++ cmd/controller/main.go
@@ -59,9 +59,6 @@ func main() {
 func reconcile(event watch.Event) error {
 	desiredExists := event.Type != watch.Deleted
 
-
-
-
 	table, ok := event.Object.(*unstructured.Unstructured)
 	if !ok {
 		return fmt.Errorf("unexpected object type: %T", event.Object)
@@ -72,7 +69,6 @@ func reconcile(event watch.Event) error {
 
 	fmt.Printf("Reconciling %s %s (event=%s)\n", namespace, resourceName, event.Type)
 
-	dbType, _, _ := unstructured.NestedString(table.Object, "spec", "type")
     project, _, _ := unstructured.NestedString(table.Object, "spec", "project")
     dataset, _, _ := unstructured.NestedString(table.Object, "spec", "dataset")
     tableName, _, _ := unstructured.NestedString(table.Object, "spec", "tableName")
@@ -88,20 +84,20 @@ func reconcile(event watch.Event) error {
         }
     }
 
-	    switch event.Type {
-    case watch.Added:
-        fmt.Printf("Creating table %s.%s (type=%s)\n", dataset, tableName, dbType)
-        return bigquery.CreateTable(project, dataset, tableName)
+	actualExists, err := bigquery.TableExists(project, dataset, tableName)
+	if err != nil {
+		return err
+	}
 
-    case watch.Deleted:
-        fmt.Printf("Deleting table %s.%s\n", dataset, tableName)
-        return bigquery.DeleteTable(project, dataset, tableName)
+	if desiredExists == actualExists {
+		fmt.Printf("In sync. desired=%v actual=%v\n", desiredExists, actualExists)
+		return nil
+	}
 
-    case watch.Modified:
-        fmt.Printf("Modified event for %s/%s (not implemented yet)\n", namespace, resourceName)
-        return nil
-
-    default:
-        return nil
-    }
+	if desiredExists {
+		fmt.Printf("Out of sync -> creating table\n")
+		return bigquery.CreateTable(project, dataset, tableName)
+	}
+	fmt.Printf("Out of sync -> deleting table\n")
+	return bigquery.DeleteTable(project, dataset, tableName)
 }
